Relocate storenode to types
